Add results of running tests without IOMMU, using the uio_pci_generic module. The net_qede driver can still get into a state of the reception getting stuck when run using the 10G bandwidth, but slowing down the rate allowed to run OK for ~30 minutes so could be a timing issue.
Code changes:
a. Add option to specify the link speed on the command line, to allow testing with the net_qede driver which fails to obtain the link speed with rte_eth_link_get_nowait() when the IOMMU isn't used.
b. Bug fix in error handling to stop segmentation fault if assertion failure occurs in the main thread.